Update initial commit

+++ src/main/java/org/example/week1/BinaryConverter.java
@@ -13,7 +13,8 @@ public class BinaryConverter {
         while( startIndex <= lastIndex ){
             char extracted = input.charAt(startIndex);
             int gottenFromExtracted = Integer.parseInt(String.valueOf(extracted));
-            total += gottenFromExtracted * Math.pow(2, (lastIndex - startIndex));
+//            total += gottenFromExtracted * Math.pow(2, (lastIndex - startIndex));
+            total += (int) (gottenFromExtracted * Math.pow(2, (lastIndex - startIndex)));
             startIndex++;
         }
         return total;
